requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
dateparser==1.2.0
frozenlist==1.4.1
idna==3.8
multidict==6.0.5
numpy==2.1.1
pandas==2.2.2
pandas_ta==0.3.14b0
pycryptodome==3.20.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.7.24
requests==2.32.3
six==1.16.0
ta==0.11.0
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
websockets==13.0.1
yarl==1.10.0
